add function to generate input noise from genomes
I added the extract_chromosome_information() function and some outcommented code in the bottom on how to run it for all data folders we have. The first folders already have their networkinput file generated, so they will be fast, the others may take a bit to generate this the first time we look at their plots, but then the filed is stored and can be used over and over again.